Implement editor-element
