+++ src/components.ts
@@ -1,5 +1,6 @@
 import "./components/home-element";
 import "./components/settings-element";
+import "./components/editor-element";
 import "./components/test-header";
 import "./components/codeEditor";
 import "./components/snippet-list";
